Pull out GivenAnActivity into its own file
It's common across any test context that requires an Activity.

Also, move the RobolectricTestUtils into that Activity as they're only
used in the context of an Activity.
