--- verticalstepper/src/test/java/com/snowble/android/widget/verticalstepper/RobolectricTestUtils.java
@@ -1,30 +1,0 @@
-package com.snowble.android.widget.verticalstepper;
-
-import android.app.Activity;
-
-import org.robolectric.Robolectric;
-
-class RobolectricTestUtils {
-
-    static VerticalStepper.LayoutParams createTestLayoutParams(Activity activity,
-                                                               int leftMargin, int topMargin,
-                                                               int rightMargin, int bottomMargin) {
-        VerticalStepper.LayoutParams lp = createTestLayoutParams(activity);
-        lp.leftMargin = leftMargin;
-        lp.topMargin = topMargin;
-        lp.rightMargin = rightMargin;
-        lp.bottomMargin = bottomMargin;
-
-        return lp;
-    }
-
-    static VerticalStepper.LayoutParams createTestLayoutParams(Activity activity) {
-        Robolectric.AttributeSetBuilder attributeSetBuilder = Robolectric.buildAttributeSet();
-        attributeSetBuilder.addAttribute(android.R.attr.layout_width, "match_parent");
-        attributeSetBuilder.addAttribute(android.R.attr.layout_height, "wrap_content");
-        attributeSetBuilder.addAttribute(R.attr.step_title, "title");
-        attributeSetBuilder.addAttribute(R.attr.step_summary, "summary");
-
-        return new VerticalStepper.LayoutParams(activity, attributeSetBuilder.build());
-    }
-}
